% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace_patterns.R
\name{attach_new_prefix}
\alias{attach_new_prefix}
\title{attach_new_prefix}
\usage{
attach_new_prefix(the_files = NULL, prefix = NULL)
}
\arguments{
\item{the_files}{character vector of file names that will receive new prefix.}

\item{prefix}{Proposed new prefix.}
}
\description{
attach proposed prefix.
}
