% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace_patterns.R
\name{create_new_prefix}
\alias{create_new_prefix}
\title{create_new_prefix}
\usage{
create_new_prefix(the_files = NULL, digits = 4)
}
\arguments{
\item{the_files}{character_vector}

\item{digits}{sprintf will nice format prefix, padding zeroes to front if
necessary.}
}
\description{
propose new prefix, min of digits
}
