% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace_patterns.R
\name{get_base_name}
\alias{get_base_name}
\title{get_base_name
 @description Remove any path information and return base filename and
 extension.
 @example
 # DT[, name := sapply(name, get_base_name)]
 @export}
\usage{
get_base_name(e)
}
\description{
get_base_name
 @description Remove any path information and return base filename and
 extension.
 @example
 # DT[, name := sapply(name, get_base_name)]
 @export
}
