% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace_patterns.R
\name{get_file_names}
\alias{get_file_names}
\title{get_file_names
 @description Returns dt, with file names and file size in MB
 @export}
\usage{
get_file_names()
}
\description{
get_file_names
 @description Returns dt, with file names and file size in MB
 @export
}
