% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_legacy_files.R
\name{get_RMD_files}
\alias{get_RMD_files}
\title{get_RMD_files}
\usage{
get_RMD_files(path = ".", pattern = NULL, recursive = FALSE)
}
\arguments{
\item{path}{Directory to query.}

\item{pattern}{regex pattern to match files}
}
\value{
full file name (with directory)
}
\description{
Given a path and pattern, return char vector of file names.
Ignores files that begin with '_'
}
\keyword{internal}
