% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace_patterns.R
\name{remove_prefix}
\alias{remove_prefix}
\title{remove_prefix}
\usage{
remove_prefix(the_files = NULL, pattern = pattern)
}
\arguments{
\item{the_files}{character vector of all possible file names to be changed.}

\item{pattern}{Select the files to change.}
}
\description{
Remove any existing prefix according to pattern
 May need multiple passes
}
