% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_legacy_files.R
\name{replace_pattern}
\alias{replace_pattern}
\title{replace_pattern}
\usage{
replace_pattern(the_files, pattern, replace)
}
\arguments{
\item{the_files}{character vector}

\item{pattern}{Regex pattern to match}

\item{replace}{Text to replace matched pattern.}
}
\description{
Given a character vector, find and replace all matching patterns.
}
\details{
This is stub.   It will replace numerous files (below) that remove
   prefix, generate new prefix and attach the new prefix.
}
\keyword{internal}
