% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace_patterns.R
\name{save_old_names}
\alias{save_old_names}
\title{save_old_names}
\usage{
save_old_names(the_files = NULL)
}
\arguments{
\item{the_files}{Unchanged character vector of file names as read from
drive.}
}
\description{
safely save copy of original files in separate environmentt
}
